lambda/embedding-lambda/requirements.txt
requests>=2.31.0
Pillow>=10.0.0
voyageai>=0.3.5
supabase>=2.0.0
